imported pymich contract loader
The pymich test folder is not shipeed with pymich on pypi. Thus the
import failed in CI. I've manually added it here until we find a better
solution. I also seems that the `ContractLoader.deploy` method was not
pushed to the pymich repo either. I've re-implemented one. Somehow I
couldn't get it under 2 parameters rather than 1 in the original test. I
also didn't what object that method was initially returning, so I
returned a straight pytezos ContractInterface object and adapted to code
test to take this into account