fix confirm txs unicity
since djwebdapp passes the level in provider.index and overides it when
starting from earliest tx to confirm, you can end up with
TezosTransactions with the same hash-nonce-counter triplet, but
different level, which throws at indexing. Hence we don't filter for
transactions with the level anymore, and only pass it when first
indexing the transaction.

to be edited with examples