implemented a small dependency mechanism for transactions
We should watch out for the performance of this algorithm. The goal is
to have a non blocking mecanism when only some transactions are aborted
amongst many to deploy. It was decided that if a dependency is aborted,
then all ascendency should be aborted as well.

since the spooler runs in his own container and is not blocking the http
requests, I made the choice to accept this costly dependency resolution
algorithm. Indeed, we need to query each dependencies and sub
dependencies which could take many requests.

Note however that I don't expect contract setups to have more that a
tree of depth > 5 and breadth > 5.

However, there might be a better way to handle this dependency resolution.