set txgroups to done when relations are created
This commit fixes the indexing of internal transactions. Until
this commit, the caller transaction was created with the `done` state,
triggering the indexer normalizers without having the internal call
relations created. This prevented access to the internal calls when
indexing caller relations.

Here, we first set all indexed transactions for a given txgroup to
`held`, preventing the indexer normalizers from catching it, then we
save all internal calls to `held` as well, creating the caller-internal
calls relations. Once this is done, we set the caller and its internal
call states to done, triggering the indexer normalizers and allowing
accesses to the internal calls from the caller transaction at normalization.