normalize caller tx & internal txs once relations are setup
We want to normalize our transactions once the caller/callee relations
are constructed in the database so that we can access internal txs when
normalizing a 'caller' transaction and vice versa.

Before this commit, we would call the normalization methods when
indexing individual transactions from the same txgroup, that is, we'd
normalize some transactions before all txs from the txgroup were indexed
preventing accessing caller/callees at normalization.

We now index all transactions from the txgroup and only then, normalize
them all in the order that they were indexed.