Override TezosCall save function to fill contract field
To keep the typing the more explicit possible, we can keep target_contract
as an attribute for the call.
But the spool function in the Tezos provider
ignore tx that doesn't have a contract filled up. So to avoid having
to fill two fields with the same value, we override the save function.