tmp: added dev docker compose.
Due to some recent update somewhere, Pytezos cannot be installed as its
dependencies never resolve and the resolver keeps backtracking, a docker
compose file is provided for `djwebdapp_ethereum` so that one can develop
from it. Indeed, the Dockerfile is based on `yourlabs/blockchain` that
has pytezos preinstalled (from when the resolver did not I presume).