feat: Separate the EthereumProvider and EthereumEventProvider
The updated version of the overridden index_init function in
EthereumEvent retrieves logs that are included between
blockchain.index_level and the head. When using an API (such as Alchemy)
to fetch these logs, there can be an instance during reindexing
where the number of logs to be retrieved exceeds expectations.
In such scenarios, the get_logs function might fail. To address this,
a retry is attempted with smaller increments between levels,
ensuring that the volume of logs does not surpass the API's limits.
Additionally, the value of these increments can be customized
in the blockchain configuration.

When indexing with logs, if the calls to index emits events as an
interesting feature is that we now in advance which level to index and
so we can prevent indexing unnecessary levels.