fix: index spooled txs.
Transactions that were spooled on blockchain that mint fast such as
Ethereum would never end up in a `done` state. The reason being that
DjWebDapp sets the level of a spooled transaction to `head` when it is
deployed, but the transaction might only be included in blocks `head+n`.
Since the `EthereumEventProvider` checkes for deployed hashes only at
the deployed level (`head`), it could not find it (since the tx was
actually included in level `head+n`).

This commit fixes this in the case where an event was emitted in the
spooled transaction. Indeed, the event retrieved when querying the block
itself includes the proper level at which the spooled transaction was
deployed and thus updates the level from `head` to `head+n`. The
transaction should then be found in the next indexing iteration.