fix: event transaction indexed as function / non indexed.
This prevents the indexed transaction (with no address) to be filtered
in `Provider.contracts` as the search filter would then contain a `None`
address. Moreover, by default, the transaction type is `contract` which
is in most cases wrong (only a contract deployment emitting a log would
be of kind `contract`). Here, we only set the transaction kind if it is
created (ie: no spooled).

Moreover, we wrap the setting of the transaction kind and the indexing
in an atomic block to ensure that all is atomic.