fix: double indexing of event-based created transactions.
The transaction triggering a log emission can be created during the log
normalization. In this case, we do not have access to the contract that
the transaction interacted with during the transaction. In such cases,
before this commit, if the blockchain had a `min_confirmation` greater
than zero, this transaction would be set as `confirm`. As such, it would
trigger a re-index in until it is confirmed and set to `done`.

During the re-indexation, the indexer looks for hashes in its database
and would call `index_call` on the hashes for transactions created in
`index_log`. However, since the transactions were searched for using
`contract.call_set`, and that the created transaction has no `contract`
attribute, the created transaction would not be found in the contract's
`call_set`.

This would result in a new transaction being created, with the same
hash resulting in 2 transactions with the same hash for the same
database to be stored. This in turn would make the indexation fail
during `Transaction.objects.update_or_create(hash=hash, ...)` in
`index_log` as the logs would be queried again until the created
transaction is set to `confirm`.